Rename a util file
